fix(bench): extract shared constants to shared_test.go
hitBinary, idleBinary, missBinary and buildTargets were only compiled
with the !userspace tag, causing build failures when running with
-tags userspace. Move them to shared_test.go (//go:build linux) so
both build configurations can see them.
